pyproject.toml
[project]
name = "gcash-nlp-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.97.0",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.9",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "protobuf>=7.34.1",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.1",
    "streamlit>=1.56.0",
    "tiktoken>=0.12.0",
    "torch>=2.11.0",
    "transformers>=5.6.2",
    "umap-learn>=0.5.12",
]
